little progress on tests

+++ index.coffee
@@ -1,44 +1,54 @@
 shell = require 'shelljs/global'
 _ = require 'lodash'
 readline = require('readline')
-rl = readline.createInterface input: process.stdin, output: process.stderr
 glob = require 'glob'
-ask = rl.question.bind rl
-finishAsking = rl.close.bind rl
-say = console.log.bind console
-log = say
-rxDPConflict = /(.*)\/(.*) \((.*) conflicted copy ([0-9\-]*)\)\.(.*)/
-rxFile = /(.*)\/(.*)\.(.*)/
+fs = require 'fs'
+
+rli = readline.createInterface input: process.stdin, output: process.stdout
+ask = rli.question.bind rli
+finishAsking = rli.close.bind rli
+log = console.log.bind console
+err = console.error.bind console
+rxDPConflict = /(.*)\/(.*) \((.*) conflicted copy ([0-9\-]*)\)(.*)/
+rxFile = /(.*)\/(.*)/
+
 
 class File
     constructor: (@path) ->
         dpFileMatch = @path.match rxDPConflict
         if (@conflicted = dpFileMatch isnt null)
-            [@path, @base, @name, @from, @date, @ext] = dpFileMatch
+            [@path, @base, @name1, @from, @date, @name2] = dpFileMatch
+            @name = @name1 + @name2
         else if not @conflicted
             regularFileMatch = @path.match rxFile
-            [@path, @base, @name, @ext] = regularFileMatch
-    matches: (otherFile) -> 
-        otherFile.name is @name and otherFile.ext is @ext and otherFile.base is @base
+            [@path, @base, @name] = regularFileMatch
+    matches: (otherFile) ->
+        otherFile.name is @name and otherFile.base is @base
+    toRegularFile: () -> path.join @base, @name
+    #TODO? toConflictedFile: (from = @from, date = @date) -> path.join @base, @name, @ext
     toString: () -> @path
 
 dpdiff =
+    ask: (cfFile) ->
+        if not fs.existsSync cfFile
+            return err "#{cfFile} doesn't exist."
+        exec ['diff', cfFile.toRegularFile(), cfFile].join(" ")
+        ask "[0] Don't do anything.\n[1] Keep non-conflicted.\n[2] Keep conflicted.\nAnswer: ", (ans) ->
+            i = parseInt ans
+            switch i
+                when 0 then say "Ok."
+                when 1 then rm nonConflictedFile
+                when 2 then rm conflictedFile
+            finishAsking()
+
     findConflictedIn: (path) ->
-        allFilePaths = glob.sync "/Users/wie/Dropbox/code/github/dpdiff/#{path}/**/*", {dot: true, nodir: true}
+        allFilePaths = glob.sync "#{path}/**/*", {dot: true, nodir: true}
         allFiles = allFilePaths.map (filePath) -> new File filePath
         [allConflicted, allNon] = _.partition allFiles, "conflicted", true
-        allConflicted.filter (cfFile) -> allNon.filter((regFile) -> cfFile.matches regFile).length > 0
-        
-main = (file1, file2) ->
-    conflictedFile = if rx.conflictedCopy.test(file1) then file1 else file2
-    nonConflictedFile = if conflictedFile is file1 then file2 else file1
-    exec ['diff', nonConflictedFile, conflictedFile].join(" ")
-    ask "[0] Don't do anything.\n[1] Keep non-conflicted.\n[2] Keep conflicted.\nAnswer: ", (ans) ->
-        i = parseInt ans
-        switch i
-            when 0 then say "Ok."
-            when 1 then rm nonConflictedFile
-            when 2 then rm conflictedFile
-        finishAsking()
+        allConflicted.filter (cfFile) -> (allNon.filter (regFile) -> cfFile.matches regFile).length > 0
 
 module.exports = dpdiff
+if require.main is module
+    path = process.argv[2]
+    conflicted = dpdiff.findConflictedIn path
+    conflicted.forEach (cfFile) -> ask cfFile
